+++ Assets/Scripts/Networking/Server/Server.cs
@@ -16,7 +16,7 @@ namespace Wheeled.Networking.Server
 
             void DisconnectedFrom(NetworkManager.Peer _peer);
 
-            void LatencyUpdated(NetworkManager.Peer _peer, int _latency);
+            void LatencyUpdated(NetworkManager.Peer _peer, float _latency);
 
             void ReceivedFrom(NetworkManager.Peer _peer, NetPacketReader _reader);
