+++ Assets/Scripts/Networking/Server/Server_Interfaces.cs
@@ -24,7 +24,7 @@ namespace Wheeled.Networking.Server
         {
         }
 
-        void NetworkManager.IEventListener.LatencyUpdated(NetworkManager.Peer _peer, int _latency)
+        void NetworkManager.IEventListener.LatencyUpdated(NetworkManager.Peer _peer, float _latency)
         {
             m_game?.LatencyUpdated(_peer, _latency);
         }
